+++ app/Http/Controllers/Partner/Api/EventController.php
@@ -38,7 +38,7 @@ final class EventController extends Controller
     {
         $validator = EventValidators::create($request->all());
         $court = $courtServices->find($validator->safe()->input('court_id'));
-        Truthy($court->user_id !== (int) Auth::id(), 'invalid operation');
+        Truthy($court->user_id !== Auth::id(), 'invalid operation');
         $event = $this->services->create(
             Auth::user(),
             $validator->safe()->except(['cords', 'days', 'times', 'tags'])
